# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SocialApp API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface UnreadCount {
  count: number;
}

export interface RegisterInput {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  displayName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  website?: string | null;
  isVerified: boolean;
  followersCount: number;
  followingCount: number;
  postsCount: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserSummary {
  id: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  isVerified: boolean;
  isFollowing: boolean;
}

export type UserProfile = User & {
  isFollowing: boolean;
  isFollowedBy: boolean;
};

export interface UpdateUserInput {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  website?: string;
}

export interface UserPage {
  users: UserSummary[];
  nextCursor?: string | null;
}

export interface Post {
  id: string;
  author: UserSummary;
  content?: string | null;
  imageUrl?: string | null;
  videoUrl?: string | null;
  location?: string | null;
  hashtags: string[];
  likesCount: number;
  commentsCount: number;
  isLiked: boolean;
  isSaved: boolean;
  createdAt: string;
}

export interface CreatePostInput {
  content?: string;
  imageUrl?: string;
  videoUrl?: string;
  location?: string;
  hashtags?: string[];
}

export interface PostPage {
  posts: Post[];
  nextCursor?: string | null;
}

export interface Comment {
  id: string;
  postId: string;
  author: UserSummary;
  content: string;
  parentId?: string | null;
  repliesCount: number;
  createdAt: string;
}

export interface CreateCommentInput {
  content: string;
  parentId?: string;
}

export interface CommentPage {
  comments: Comment[];
  nextCursor?: string | null;
}

export type StoryMediaType =
  (typeof StoryMediaType)[keyof typeof StoryMediaType];

export const StoryMediaType = {
  image: "image",
  video: "video",
} as const;

export interface Story {
  id: string;
  userId: string;
  mediaUrl: string;
  mediaType: StoryMediaType;
  /** Duration in seconds */
  duration: number;
  viewsCount: number;
  isViewed: boolean;
  createdAt: string;
  expiresAt: string;
}

export interface StoryGroup {
  user: UserSummary;
  stories: Story[];
  hasUnviewed: boolean;
}

export type CreateStoryInputMediaType =
  (typeof CreateStoryInputMediaType)[keyof typeof CreateStoryInputMediaType];

export const CreateStoryInputMediaType = {
  image: "image",
  video: "video",
} as const;

export interface CreateStoryInput {
  mediaUrl: string;
  mediaType: CreateStoryInputMediaType;
  duration?: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  like: "like",
  comment: "comment",
  follow: "follow",
  mention: "mention",
  reply: "reply",
} as const;

export interface Notification {
  id: string;
  type: NotificationType;
  fromUser: UserSummary;
  post?: Post | null;
  commentContent?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationPage {
  notifications: Notification[];
  nextCursor?: string | null;
}

export interface Message {
  id: string;
  conversationId: string;
  senderId: string;
  content?: string | null;
  mediaUrl?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface Conversation {
  id: string;
  participant: UserSummary;
  lastMessage?: Message | null;
  unreadCount: number;
  updatedAt: string;
}

export interface MessagePage {
  messages: Message[];
  nextCursor?: string | null;
}

export interface SendMessageInput {
  content?: string;
  mediaUrl?: string;
}

export interface StartConversationInput {
  userId: string;
}

export interface Hashtag {
  name: string;
  postCount: number;
}

export interface SearchResults {
  users: UserSummary[];
  posts: Post[];
  hashtags: Hashtag[];
}

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Conflict
 */
export type ConflictResponse = ErrorResponse;

export type CursorParameter = string;

export type LimitParameter = number;

export type GetUserPostsParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetFollowersParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetFollowingParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetCommentsParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetFeedParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetExplorePostsParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetNotificationsParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type GetMessagesParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};

export type SearchParams = {
  q: string;
  type?: SearchType;
};

export type SearchType = (typeof SearchType)[keyof typeof SearchType];

export const SearchType = {
  users: "users",
  posts: "posts",
  hashtags: "hashtags",
  all: "all",
} as const;

export type GetSavedPostsParams = {
  cursor?: CursorParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: LimitParameter;
};