# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialApp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Users ───────────────────────────────────────────────────────────────────
...
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [likes]
      summary: Like a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/postId"
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unlikePost
      tags: [likes]
      summary: Unlike a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/postId"
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments on a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/postId"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentPage"
    post:
      operationId: createComment
      tags: [comments]
      summary: Comment on a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/postId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/commentId"
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Feed ────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    postId:
      name: postId
      in: path
      required: true
      schema:
        type: string
...
    commentId:
      name: commentId
      in: path
      required: true
      schema:
        type: string
...
    cursor:
      name: cursor
      in: query
      schema:
        type: string
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    RegisterInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
      required: [username, email, password, displayName]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # ── Users ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        isVerified:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, displayName, isVerified, followersCount, followingCount, postsCount, createdAt]
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        isVerified:
          type: boolean
        isFollowing:
          type: boolean
      required: [id, username, displayName, isVerified, isFollowing]
...
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        author:
          $ref: "#/components/schemas/UserSummary"
        content:
          type: string
        parentId:
          type: string
          nullable: true
        repliesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, postId, author, content, repliesCount, createdAt]
    CreateCommentInput:
      type: object
      properties:
        content:
          type: string
        parentId:
          type: string
      required: [content]
    CommentPage:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        nextCursor:
          type: string
          nullable: true
      required: [comments]

    # ── Stories ───────────────────────────────────────────────────────────────
...